Fix write on read bug with lww+bitcask
Fixes issue #334.
Reading a value written into a bucket with last write wins set to true
when using a non index supporting backend (bitcask) was causing
read-repairs writes.  The problem is the logic was causing all replica
writing nodes to increment the vclock, instead of just incrementing in
the coordinating node and forwarding to replicas as normally happens.