Change AAE hashtrees to buffer/batch deletes
AAE already buffers writes and submits them as a single write-batch to
LevelDB for better performance. Since deletes are also valid batch
operations, this change buffers deletes alongside writes and submits
all operations as a single write-batch as necessary.

This change also fixes a potential bug in the previous approach that
could reorder a write to happen after a delete, even if the delete
occured after the write -- thus reviving a dead hash.