Adding bad_crc specifically on read before write
Although plenty of other situations may arise where an error upon
reading should simply allow the write to happen, we're going to wait to
standardize backend responses properly before trying to fix those.
This is just to be prudent until the consequences of doing that are
better understood. This makes the fix specific to bitcask bad CRC
errors and nothing more.
Standardization will hopefully happen in issue #496