Feature request, optionally return indexed values with keys
When a 2i query's results are a list of keys. This change adds
an optional request parameter to pb and http that, when set, will
cause pairs of `key:index_value` to be returned as the result of
a 2i query.

Has no effect on MapReduce. Uses capabilities for rolling upgrade
situations.