make multibackend work as advertised
The multibackend performs set intersection to determine intersecting capabilities
of all sub-backends but it used the empty set as Acc0, which resulted in the empty set
always. This addresses this issue by using ordsets:intersection/1 instead.

Note: this has the side-effect of making the multi-backend index capable if all
sub-backends support indexing (e.g. some mix of memory and leveldb)