Adds X-Riak-Deleted where missing
In the case of a single object being returned which is a
tombstone and the case where a single sibling has been
requested via a vtag and it is a tombstone, the
X-Riak-Deleted header will now be added and
returned.

Ideally both these would return 204 rather than 404 and 200
respectively but doing so would have breaking consequences
for clients.