Disable get/put monitors when using FSM overload protection
The get/put monitors are no longer needed when using sidejob_supervisors
to manage get/put FSMs. This commit disables the use of get/put monitors
when overload protection is enabled, as well as changes the existing
`puts_active` and `gets_active` API to use either the get/put monitor
data or sidejob_supervisor data depending on which option is being used.