Handle worker errors and timeout on 2i reformat
Before this fix, I noticed the reformat basically hang forever with a
crashed worker error once. This fix handles the error message from the
vnode worker and adds a big 2 minute timeout to make sure we don't get
stuck. In reality, it's mostly harmless, but annoying to leave a process
hanging forever like that.