First pass at new riak-admin test cluster functionality
Do a read/write cycle for each node reported in [node()|nodes()]. So,
this will not include down nodes. Also this is a riak_client get/put,
still up in the air if I want to do a vnode put or if a client put is
good enugh(taking into account qourem, etc).