Avoid t2b/b2t between vnode and get FSM
This is a port of Krestenkrab's PR :
https://github.com/basho/riak_kv/pull/83
Avoids doing a binary to term on an object that is sent from a vnode to
the get FSM requesting it. This means that the FSM will have to do it on
that side, but it avoids the implicit term_to_binary/binary_to_term when
the object is sent through disterl.  Some tests using a memory backend
on OSX show a decent improvement (~5% in throughput and more in
latency).
The get_fsm fake vnode had to be updated to support the get binary
request.