Added get_fsm get binary obj capability
The get_fsm will only start  requesting binary objects when fully
upgraded, not in a mixed cluster. However, the downgrade path could be
as painful as with other things using capabilities. That is, for a short
time during a downgrade while capabilities are propagated, get requests
to certain vnodes could fail. If kept to a node at a time though, it
should be transparent to requests without lost replicas.