Fix orphan stat repair process
A customer had a good number of riak_kv_stat restarts due to problems in
folsom as of yet undetermined. The number of processes was observed to
balloon in size. The stat repair process was created every time a
riak_kv_stat process was restarted. riak_kv_stat monitored this process,
but this process did not monitor riak_kv_stat, so when that went away,
this process stayed around forever blocked waiting on its msg box.
This change ensures the process monitors riak_kv_stat so it goes down
with it.