Added warning about removing mutator code.
Added a similar warning in 4 places concerning the same thing. If an object
is stored using a given mutator, that same mutator is used when the object
is retreived. If that code isn't available or has changed in a backwards
incompatible way, one can expect to have a bad time.