add option to filter keys that are no longer in preflist during fold
In cases like reducing n-value and ring resizing Riak may leave additional
replicas on a partition that no longer is a member of a preflist for the
replica's key. Riak does not (yet) clean up this data, and even if it did,
there would be a window between the time the data was considered "straggling"
and when it was deleted. To prevent this a filter for user-facing folds (not handoff
folds -- via FOLD_REQ) has been added to check whether or not the key being folded
over has the partition as a member of its preflist.

This change, also, does not affect bucket folds because during these
folds the keys are not exposed and it would be expensive to verify if
at least one key in the bucket had the partition as a member of its
preflist.