Add 2i term regex filter
Adds a regular expression filter for secondary index terms. With this
you can scan a range for a binary index from, say A through Z, but only
use items that match terms containing "AA".
This filter does not work with integer queries, but it does work with
$key queries.

Regex compilation: currently it's done in the vnode, as different nodes
might have different versions of the regular expression library.
The PB and HTTP will compile the regular expression to validate it and
will return an error if it is invalid.