Make 2i sort default configurable
This makes the default be specified in only one place and gives people
with migration problems a way to change it without modifying their code
which has been a point of contention with this sorting business.

Cleaned up inconsistent use of sort/order. Let's go with sort flag for
now.