Keep a per vnode counter and use it to create the "fresh" vclock
When a local "notfound" occurs on the get for a co-ord put, use
the vnode counter to create a fresh vclock and merge it with the
received objects vclock. See riak_kv 679 for details.

Also, needs riak_core/gh679
WIP, considering just keeping a vclock at the vnode rather than
constructing a fresh one when needed.