Fix consistent_object mixed-mode failure.
Ensure that if we attempt to call consistent_object on a remote node
and get a undefined function error back, this most likely means that
the node contacted is running an earlier version of Riak, in which
case strong consistency is not available.

It appears that in addition to this, we probably should have a
capability for this.