Fix vnode sending > max_results items for 2i query
This change sends the max_results value in the 2i query to the vnode,
which then may be used as the batch size if < than the default batch
size.  Before, a batch would be sent with possibly more items than the
index FSM would ever send back to the client.