Add yz_stat exception to xref analysis
Running xref complains about undefined calls to yz_stat because
Yokozuna is not a riak_kv dependency.

Since we already ignore yz_kv calls, this commit adds an exception for
yz_stat calls as well.

In the future, we'll either make Yokozuna a proper sub-dependency of
riak_kv, or add support to dynamically registered stat callbacks for
stat aggregation (eg. for /stats).