Remove unused key_counts/0|1
Pair-programmed with @reiddraper

Dialyzer found a bug with this function, and it appears to be completely unused
in Riak. It was introduced in 39343fab. In the creation of the record, ref is
set to undefined, which will crash the underlying call to Bitcask when it tries
to grab information of of the process dictionary.