Create a one-arity build_filter function
When called with `undefined` as the last argument, build_filter/3 was
ignoring its first argument (it still does for backward compatibility).
Because we were providing it with a dummy 'all' argument, dialyzer was
complaining. Since we know at the call-site whether we're calling the
function with `undefined` or not, we can just use the new one-arity
version.

Fixes #923