Bound some EQC tests by time not count
With the KV tests taking about 30 minutes on the buildbot builders, this
should help reduce the testing time a bit. I've also fixed up some cases
where we thought the eunit timoeout was in milliseconds (its not,
anymore anyway). We do, however, need to be careful that we don't make
the tests significantly slower, as this may be easy to hide, and then we
might only run a few dozen tests in the testing time.