Correct discrepancy between do_get_term and case clause in do_delete
This was causing node shutdown when running the Java client
integration tests. On a small scale, this would be fine (one vnode
crashing would start back up), but in massive-delete scenarios entire
clusters could go down. We should heed the TODO message in do_get_term
and handle this holistically, but for now this should prevent the
vnode from crashing.