Ignore ignore file numbers, but consider count
The bulk of the code deals with the dialyzer.ignore-warnings file
which contains strings to ignore if output by dialyzer.
Typically the strings include line numbers. Using them exactly is hard
to maintain as the code changes. This approach instead ignores the line
numbers, but takes into account the number of times a string is listed
for a given file (See the sort + uniq -c commands, which add the count
as a prefix). So if one string is listed once, for example, and it
appears twice in the warnings for that file, the user is alerted.
It is possible but unlikely that this approach could mask a warning if one
ignored warning is removed and two warnings of the same kind appear in the
file, for example. But it is a trade-off that seems worth it.