Add synctree_path callback to ensemble backend
This commit adds the new synctree_path callback to the K/V ensemble
backend. This callback is used to override how riak_ensemble maps
peer synctrees to on-disk LevelDB instances.

This commit implements the callback using a M:N mapping, where all
peers that correspond to a single vnode map to a common LevelDB
instance. This makes the number of synctree LevelDB instances scale
with the size of the ring, rather than the number of distinct preflists.

This is the same approach used for AAE in hashtree.erl.

Requiring fewer instances makes better use of file handles, write buffers,
and internal LevelDB caches.