Add RangeReg as a new supported datatype
Rangereg is a register for integers that tracks the largest,
the smallest, the first (by timestamp) and the last (by timestamp)
integers assigned into it.

The new type is called "rangereg" inside the codebase, and can be
activated just like any other bucket-type datatype. It exposes
the exact same statistics as the existing 2.0 datatypes.